Add and remove screening helper fields from output. Fix bugs
Added 2 fields for hatch date, calculated from 1st observed stage 1 and 1st observed stage 4.
Removed helper fields active.last.date, max.stage, num.spp from output; these not helpfull.

Changed calculation of days.short.of.fledging to be calculated based on latest date with chicks instead of latest date unguarded, also to be calculated for any nest observed in stage 2 or higher, instead of stage 4 or higher: to account for long stage 3 and so this is calculated for nests that fail shortly after hatching

Changed determination of spp.max.date to be for any nest, not just active nests; ensures a final check with all nests innactive is accounted for.